refactor(vue): cleanup files
