+++ code/renderers/vue3/src/docgen/component-meta/follow-re-export.ts
