+++ code/renderers/vue3/src/story-docs/print/print-h.ts
@@ -6,8 +6,8 @@ import {
   type ImportBinding,
 } from 'storybook/internal/csf-tools';
 
-import { classifyArg, type ClassifiedSlotArg, type VueDocgenArgInfo } from './classify-args.ts';
-import { isFunctionExpression, printValue } from './classify-value.ts';
+import { classifyArg, type ClassifiedSlotArg, type VueDocgenArgInfo } from '../classify/args.ts';
+import { isFunctionExpression, printValue } from '../shared/values.ts';
 import {
   escapeTextContent,
   formatRenderedProp,
@@ -19,7 +19,7 @@ import {
   renderSlotContent,
   wrapSlotContent,
   type RenderContext,
-} from './render-primitives.ts';
+} from '../shared/primitives.ts';
 
 export interface PrintHInput {
   /** Render-function expression to print as template markup. */
