+++ code/renderers/vue3/src/story-docs/print/slot-arg.ts
@@ -1,11 +1,16 @@
 import { unwrapExpression, type ImportBinding } from 'storybook/internal/csf-tools';
 
-import type { ClassifiedSlotArg } from './classify-args.ts';
-import { isFunctionExpression } from './classify-value.ts';
+import type { ClassifiedSlotArg } from '../classify/args.ts';
+import { isFunctionExpression } from '../shared/values.ts';
 import { printHFragment } from './print-h.ts';
-import { renderSlotContent, type RenderContext } from './render-primitives.ts';
+import { renderSlotContent, type RenderContext } from '../shared/primitives.ts';
 
-/** Slot children for one classified slot arg, or undefined when a function slot cannot render. */
+/**
+ * Slot children for one classified slot arg, or undefined when a function slot cannot render.
+ *
+ * Separate from `../shared/primitives.ts` because a function slot prints through `./print-h.ts`,
+ * which the primitives themselves back.
+ */
 export function renderSlotArgContent(
   arg: ClassifiedSlotArg,
   ctx: RenderContext,
